test: split admin instance update tests
