+++ crates/dcc-mcp-gateway/src/gateway/admin/tests.rs
@@ -30,9 +30,6 @@ mod admin_tests {
     /// point at a non-existent directory for the duration of the request.
     static API_LOGS_ENV_LOCK: Mutex<()> = Mutex::new(());
 
-    /// Update staging tests mutate platform data-dir env vars.
-    static UPDATE_ENV_LOCK: Mutex<()> = Mutex::new(());
-
     struct ScopedNoDiskLogsDir {
         previous: Option<String>,
     }
@@ -103,11 +100,6 @@ mod admin_tests {
         previous: Vec<(&'static str, Option<String>)>,
     }
 
-    struct ScopedUpdateDataDir {
-        previous: Vec<(&'static str, Option<String>)>,
-        _dir: tempfile::TempDir,
-    }
-
     impl ScopedIntegrationEnv {
         fn new(values: &[(&'static str, Option<&str>)]) -> Self {
             const KEYS: &[&str] = &[
@@ -145,65 +137,6 @@ mod admin_tests {
     }
 
     impl Drop for ScopedIntegrationEnv {
-        fn drop(&mut self) {
-            // SAFETY: same as `new` — guarded by the test mutex.
-            unsafe {
-                for (key, value) in &self.previous {
-                    match value {
-                        Some(value) => std::env::set_var(key, value),
-                        None => std::env::remove_var(key),
-                    }
-                }
-            }
-        }
-    }
-
-    impl ScopedUpdateDataDir {
-        fn new() -> Self {
-            const KEYS: &[&str] = &["APPDATA", "XDG_DATA_HOME", "HOME"];
-            let previous = KEYS
-                .iter()
-                .map(|key| (*key, std::env::var(key).ok()))
-                .collect::<Vec<_>>();
-            let dir = tempfile::tempdir().unwrap();
-            let p = dir.path().to_string_lossy().to_string();
-            // SAFETY: tests using these vars are serialized by `UPDATE_ENV_LOCK`.
-            unsafe {
-                std::env::set_var("APPDATA", &p);
-                std::env::set_var("XDG_DATA_HOME", &p);
-                std::env::set_var("HOME", &p);
-            }
-            Self {
-                previous,
-                _dir: dir,
-            }
-        }
-
-        fn root(&self) -> &std::path::Path {
-            self._dir.path()
-        }
-
-        fn pending_marker(&self, binary_name: &str) -> std::path::PathBuf {
-            #[cfg(target_os = "macos")]
-            {
-                self.root()
-                    .join("Library")
-                    .join("Application Support")
-                    .join("update")
-                    .join(binary_name)
-                    .join("pending.marker")
-            }
-            #[cfg(not(target_os = "macos"))]
-            {
-                self.root()
-                    .join("update")
-                    .join(binary_name)
-                    .join("pending.marker")
-            }
-        }
-    }
-
-    impl Drop for ScopedUpdateDataDir {
         fn drop(&mut self) {
             // SAFETY: same as `new` — guarded by the test mutex.
             unsafe {
@@ -305,24 +238,6 @@ mod admin_tests {
             .oneshot(
                 Request::builder()
                     .method("PUT")
-                    .uri(uri)
-                    .header(header::CONTENT_TYPE, "application/json")
-                    .body(axum::body::Body::from(payload.to_string()))
-                    .unwrap(),
-            )
-            .await
-            .unwrap();
-        let status = resp.status();
-        let bytes = to_bytes(resp.into_body(), 1024 * 1024).await.unwrap();
-        let json: Value = serde_json::from_slice(&bytes).unwrap_or(Value::Null);
-        (status, json)
-    }
-
-    async fn post_json(router: Router, uri: &str, payload: Value) -> (StatusCode, Value) {
-        let resp = router
-            .oneshot(
-                Request::builder()
-                    .method("POST")
                     .uri(uri)
                     .header(header::CONTENT_TYPE, "application/json")
                     .body(axum::body::Body::from(payload.to_string()))
@@ -515,103 +430,6 @@ filters:
                 .await;
         });
         (port, tx)
-    }
-
-    async fn spawn_update_manifest(manifest: Value) -> (String, oneshot::Sender<()>) {
-        let app = Router::new().route(
-            "/manifest.json",
-            axum::routing::get(move || {
-                let manifest = manifest.clone();
-                async move { axum::Json(manifest) }
-            }),
-        );
-        let listener = tokio::net::TcpListener::bind("127.0.0.1:0").await.unwrap();
-        let url = format!(
-            "http://127.0.0.1:{}/manifest.json",
-            listener.local_addr().unwrap().port()
-        );
-        let (tx, rx) = oneshot::channel::<()>();
-        tokio::spawn(async move {
-            let _ = axum::serve(listener, app)
-                .with_graceful_shutdown(async {
-                    let _ = rx.await;
-                })
-                .await;
-        });
-        (url, tx)
-    }
-
-    async fn spawn_update_manifest_with_binary(
-        version: &'static str,
-        binary_body: &'static [u8],
-    ) -> (String, oneshot::Sender<()>) {
-        let listener = tokio::net::TcpListener::bind("127.0.0.1:0").await.unwrap();
-        let port = listener.local_addr().unwrap().port();
-        let download_url = format!("http://127.0.0.1:{port}/dcc-mcp-server.bin");
-        let manifest = json!({
-            "dcc-mcp-server": {
-                "version": version,
-                "url": download_url,
-                "sha256": null,
-                "release_notes": "Server update"
-            }
-        });
-        let app = Router::new()
-            .route(
-                "/manifest.json",
-                axum::routing::get(move || {
-                    let manifest = manifest.clone();
-                    async move { axum::Json(manifest) }
-                }),
-            )
-            .route(
-                "/dcc-mcp-server.bin",
-                axum::routing::get(move || {
-                    let body = binary_body.to_vec();
-                    async move { ([(header::CONTENT_TYPE, "application/octet-stream")], body) }
-                }),
-            );
-        let url = format!("http://127.0.0.1:{port}/manifest.json");
-        let (tx, rx) = oneshot::channel::<()>();
-        tokio::spawn(async move {
-            let _ = axum::serve(listener, app)
-                .with_graceful_shutdown(async {
-                    let _ = rx.await;
-                })
-                .await;
-        });
-        (url, tx)
-    }
-
-    async fn spawn_update_manifest_response(
-        status: StatusCode,
-        content_type: &'static str,
-        body: &'static str,
-    ) -> (String, oneshot::Sender<()>) {
-        let app = Router::new().route(
-            "/manifest.json",
-            axum::routing::get(move || async move {
-                axum::response::Response::builder()
-                    .status(status)
-                    .header(header::CONTENT_TYPE, content_type)
-                    .body(axum::body::Body::from(body))
-                    .unwrap()
-            }),
-        );
-        let listener = tokio::net::TcpListener::bind("127.0.0.1:0").await.unwrap();
-        let url = format!(
-            "http://127.0.0.1:{}/manifest.json",
-            listener.local_addr().unwrap().port()
-        );
-        let (tx, rx) = oneshot::channel::<()>();
-        tokio::spawn(async move {
-            let _ = axum::serve(listener, app)
-                .with_graceful_shutdown(async {
-                    let _ = rx.await;
-                })
-                .await;
-        });
-        (url, tx)
     }
 
     async fn spawn_skill_detail_backend(hits: Value, detail: Value) -> (u16, oneshot::Sender<()>) {
@@ -2631,261 +2449,6 @@ webhooks:
         assert_eq!(workers[0]["dispatch_status"], "unavailable");
         assert_eq!(workers[0]["dispatch_ready"], false);
         assert_eq!(body["summary"]["unhealthy"].as_u64(), Some(1));
-    }
-
-    #[tokio::test]
-    async fn test_admin_instance_update_reports_missing_manifest_config() {
-        let gs = make_gateway_state();
-        let instance_id = {
-            let reg = gs.registry.write().await;
-            let entry = make_service_entry("maya", "127.0.0.1", 18813, Some(4242));
-            let instance_id = entry.instance_id.to_string();
-            reg.register(entry).unwrap();
-            instance_id
-        };
-
-        let state = AdminState::new(gs);
-        let router = build_admin_router(state);
-        let (status, body) = post_json(
-            router,
-            &format!("/api/instances/{instance_id}/update"),
-            json!({ "apply": true, "current_version": "0.18.0" }),
-        )
-        .await;
-
-        assert_eq!(status, StatusCode::NOT_IMPLEMENTED);
-        assert_eq!(body["status"], "not_configured");
-        assert_eq!(body["binary_name"], "dcc-mcp-server");
-        assert_eq!(body["current_version"], "0.18.0");
-        assert_eq!(body["current_version_source"], "request");
-        assert_eq!(body["requires_restart"], false);
-    }
-
-    #[tokio::test]
-    async fn test_admin_instance_update_requires_binary_version_for_non_server_binary() {
-        let gs = make_gateway_state();
-        let instance_id = {
-            let reg = gs.registry.write().await;
-            let entry = make_service_entry("maya", "127.0.0.1", 18813, Some(4242));
-            let instance_id = entry.instance_id.to_string();
-            reg.register(entry).unwrap();
-            instance_id
-        };
-
-        let state = AdminState::new(gs);
-        let router = build_admin_router(state);
-        let (status, body) = post_json(
-            router,
-            &format!("/api/instances/{instance_id}/update"),
-            json!({ "binary": "dcc-mcp-cli", "apply": false }),
-        )
-        .await;
-
-        assert_eq!(status, StatusCode::BAD_REQUEST);
-        assert_eq!(body["status"], "version_required");
-        assert_eq!(body["error"], "current_version_required");
-        assert_eq!(body["binary_name"], "dcc-mcp-cli");
-        assert_eq!(body["update_available"], false);
-    }
-
-    #[tokio::test]
-    async fn test_admin_instance_update_checks_manifest_without_manual_cli() {
-        let (manifest_url, shutdown) = spawn_update_manifest(json!({
-            "dcc-mcp-server": {
-                "version": env!("CARGO_PKG_VERSION"),
-                "url": null,
-                "sha256": null,
-                "release_notes": "Already current"
-            }
-        }))
-        .await;
-
-        let mut gs = make_gateway_state();
-        gs.update_manifest_url = Some(manifest_url);
-        let instance_id = {
-            let reg = gs.registry.write().await;
-            let entry = make_service_entry("maya", "127.0.0.1", 18813, Some(4242));
-            let instance_id = entry.instance_id.to_string();
-            reg.register(entry).unwrap();
-            instance_id
-        };
-
-        let state = AdminState::new(gs);
-        let router = build_admin_router(state);
-        let (status, body) = post_json(
-            router,
-            &format!("/api/instances/{instance_id}/update"),
-            json!({ "apply": true, "current_version": env!("CARGO_PKG_VERSION") }),
-        )
-        .await;
-        let _ = shutdown.send(());
-
-        assert_eq!(status, StatusCode::OK);
-        assert_eq!(body["status"], "up_to_date");
-        assert_eq!(body["update_available"], false);
-        assert_eq!(body["current_version"], env!("CARGO_PKG_VERSION"));
-        assert_eq!(body["current_version_source"], "request");
-        assert_eq!(body["latest_version"], env!("CARGO_PKG_VERSION"));
-        assert_eq!(body["requires_restart"], false);
-    }
-
-    #[tokio::test]
-    async fn test_admin_instance_update_reports_missing_download_url() {
-        let (manifest_url, shutdown) = spawn_update_manifest(json!({
-            "dcc-mcp-server": {
-                "version": "999.0.0",
-                "url": null,
-                "sha256": null,
-                "release_notes": "Update metadata without a downloadable binary"
-            }
-        }))
-        .await;
-
-        let mut gs = make_gateway_state();
-        gs.update_manifest_url = Some(manifest_url);
-        let instance_id = {
-            let reg = gs.registry.write().await;
-            let entry = make_service_entry("maya", "127.0.0.1", 18813, Some(4242));
-            let instance_id = entry.instance_id.to_string();
-            reg.register(entry).unwrap();
-            instance_id
-        };
-
-        let state = AdminState::new(gs);
-        let router = build_admin_router(state);
-        let (status, body) = post_json(
-            router,
-            &format!("/api/instances/{instance_id}/update"),
-            json!({ "apply": true, "current_version": "0.18.0" }),
-        )
-        .await;
-        let _ = shutdown.send(());
-
-        assert_eq!(status, StatusCode::NOT_FOUND);
-        assert_eq!(body["status"], "download_failed");
-        assert_eq!(body["error"], "download_url_not_configured");
-        assert_eq!(body["binary_name"], "dcc-mcp-server");
-        assert_eq!(body["current_version"], "0.18.0");
-        assert_eq!(body["current_version_source"], "request");
-        assert_eq!(body["latest_version"], "999.0.0");
-        assert_eq!(body["update_available"], true);
-        assert_eq!(body["requires_restart"], false);
-    }
-
-    #[tokio::test]
-    async fn test_admin_instance_update_can_check_without_staging() {
-        let (manifest_url, shutdown) =
-            spawn_update_manifest_with_binary("999.0.0", b"server-binary").await;
-
-        let mut gs = make_gateway_state();
-        gs.update_manifest_url = Some(manifest_url);
-        let instance_id = {
-            let reg = gs.registry.write().await;
-            let entry = make_service_entry("maya", "127.0.0.1", 18813, Some(4242));
-            let instance_id = entry.instance_id.to_string();
-            reg.register(entry).unwrap();
-            instance_id
-        };
-
-        let state = AdminState::new(gs);
-        let router = build_admin_router(state);
-        let (status, body) = post_json(
-            router,
-            &format!("/api/instances/{instance_id}/update"),
-            json!({ "apply": false, "current_version": "0.18.0" }),
-        )
-        .await;
-        let _ = shutdown.send(());
-
-        assert_eq!(status, StatusCode::OK);
-        assert_eq!(body["status"], "available");
-        assert_eq!(body["binary_name"], "dcc-mcp-server");
-        assert_eq!(body["current_version"], "0.18.0");
-        assert_eq!(body["current_version_source"], "request");
-        assert_eq!(body["latest_version"], "999.0.0");
-        assert_eq!(body["update_available"], true);
-        assert_eq!(body["requires_restart"], false);
-    }
-
-    #[tokio::test]
-    async fn test_admin_instance_update_stages_server_binary() {
-        let _guard = UPDATE_ENV_LOCK.lock();
-        let data_dir = ScopedUpdateDataDir::new();
-        let (manifest_url, shutdown) =
-            spawn_update_manifest_with_binary("999.0.0", b"server-binary").await;
-
-        let mut gs = make_gateway_state();
-        gs.update_manifest_url = Some(manifest_url);
-        let instance_id = {
-            let reg = gs.registry.write().await;
-            let entry = make_service_entry("maya", "127.0.0.1", 18813, Some(4242));
-            let instance_id = entry.instance_id.to_string();
-            reg.register(entry).unwrap();
-            instance_id
-        };
-
-        let state = AdminState::new(gs);
-        let router = build_admin_router(state);
-        let (status, body) = post_json(
-            router,
-            &format!("/api/instances/{instance_id}/update"),
-            json!({ "apply": true, "current_version": "0.18.0" }),
-        )
-        .await;
-        let _ = shutdown.send(());
-
-        assert_eq!(status, StatusCode::OK);
-        assert_eq!(body["status"], "staged");
-        assert_eq!(body["binary_name"], "dcc-mcp-server");
-        assert_eq!(body["current_version"], "0.18.0");
-        assert_eq!(body["current_version_source"], "request");
-        assert_eq!(body["latest_version"], "999.0.0");
-        assert_eq!(body["update_available"], true);
-        assert_eq!(body["requires_restart"], true);
-        assert!(
-            data_dir.pending_marker("dcc-mcp-server").exists(),
-            "staging should write the pending update marker"
-        );
-    }
-
-    #[tokio::test]
-    async fn test_admin_instance_update_reports_manifest_http_error() {
-        let (manifest_url, shutdown) = spawn_update_manifest_response(
-            StatusCode::NOT_FOUND,
-            "text/html",
-            "<!doctype html><title>missing</title>",
-        )
-        .await;
-
-        let mut gs = make_gateway_state();
-        gs.update_manifest_url = Some(manifest_url);
-        let instance_id = {
-            let reg = gs.registry.write().await;
-            let entry = make_service_entry("maya", "127.0.0.1", 18813, Some(4242));
-            let instance_id = entry.instance_id.to_string();
-            reg.register(entry).unwrap();
-            instance_id
-        };
-
-        let state = AdminState::new(gs);
-        let router = build_admin_router(state);
-        let (status, body) = post_json(
-            router,
-            &format!("/api/instances/{instance_id}/update"),
-            json!({ "apply": true, "current_version": "0.18.0" }),
-        )
-        .await;
-        let _ = shutdown.send(());
-
-        assert_eq!(status, StatusCode::BAD_GATEWAY);
-        assert_eq!(body["status"], "manifest_error");
-        assert_eq!(body["binary_name"], "dcc-mcp-server");
-        assert!(
-            body["message"]
-                .as_str()
-                .is_some_and(|detail| detail.contains("404"))
-        );
-        assert_eq!(body["requires_restart"], false);
     }
 
     #[tokio::test]
